# backend/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoAppAPI",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/users": {
            "get": {
                "description": "ユーザーIDから取得する",
                "tags": [
                    "user"
                ],
                "summary": "ユーザーを返す",
                "parameters": [
                    {
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "ユーザーを作成する",
                "tags": [
                    "user"
                ],
                "summary": "ユーザーを作成する",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/model.User"
                            }
                        }
                    },
                    "description": "ユーザー情報",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "ユーザーID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "schemas": {
            "model.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
            "model.User": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            }
        }
    }
}